refactor: restructure the objc compatibility classes
