+++ Sources/RudderStackAnalytics/ObjC/Base/ObjCSessionConfiguration.swift
